merge

+++ src/main/java/org/team2/domain/OpVO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class OpVO {
-    private Long op_seq;
-    private Long order_seq;
-    private Long product_seq;
+    private int op_seq;
+    private int order_seq;
+    private int product_seq;
     private Long op_count;
 }
